requirements.txt
fastapi
uvicorn
sqlalchemy
joblib
psycopg2-binary
pytest
scikit-learn
httpx
python-dotenv